moved jsx to separate file + validation during submit
